# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EAMS API",
    "description": "Events Attendance Monitoring System — attendance capture and administration.\n\n**This document supersedes `docs/api/attendance-contract-handoff.md` as the contract.**\nIt is generated from the source, so the outcome tokens, status codes and payload shapes\nbelow are what the running build actually does.\n\n- **Base path** `/api/v1`. JSON in and out, camelCase field names.\n- **All timestamps are UTC**, ISO 8601, with an explicit `Z`. A bare or offset-bearing\n  local time misjudges the Present/Late boundary by the server's offset.\n- **Errors are RFC 7807** `application/problem+json`. Every body carries a `traceId` to\n  quote back and, on the attendance surface, a stable `code` to branch on — never the\n  human-readable `title` or `detail`, which are reworded freely.\n- **Capture endpoints require a device key** (`DeviceKey` below). Everything else is open\n  for now: authentication for human users is Phase 6, and until it lands this API must\n  stay on a local or trusted network.\n\nSee the `TapOutcomeCode`, `ManualOutcomeCode` and `LiveOutcomeCode` schemas for the frozen\ntoken tables.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/events/{id}/attendees/students/{studentId}": {
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "`DELETE /events/{id}/attendees/students/{studentId}` — detach one individually-attached\r\n             student from the event's audience.",
        "description": "Idempotent, on the same reasoning as M:EAMS.Api.Controllers.EventsController.DetachGroup(System.Guid,System.Guid,System.Threading.CancellationToken): detaching a student who is not\r\nattached is `204`, not `404`, so a retry is safe. A missing <em>event</em> is still\r\n`404` — that one is named by the URL.\r\n            \r\n\r\n\r\nSpelled out rather than left as `<inheritdoc>`: Swashbuckle does not resolve that tag,\r\nso this operation shipped with an empty description while looking documented in the source.\r\n",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "The student to detach.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Detached, or was not attached."
          },
          "404": {
            "description": "No such event.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The event's status does not allow an audience change.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "`GET /students` — the roster, every filter optional.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Matches student number or name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "course",
            "in": "query",
            "description": "<b>ADR-001 D-2 derived display cache.</b> A student can sit in several sections at once — twelve\r\n            of fifty-two in the real roster do — so this filter cannot represent the truth. It exists for the\r\n            admin grid; do not build reporting on it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "`Active`, `Inactive` or `Graduated`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number, default 1. Out-of-range values are clamped, never refused; the response\r\nechoes the page actually served.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 50, maximum 200 — a larger value is clamped to the maximum and the\r\nresponse says so in its own `pageSize`.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matching students, possibly empty.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDtoPagedResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "§6.2 `POST /students` — manual roster entry, alongside the §10 bulk import.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentWriteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentWriteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "`GET /students/{id}` — one student by primary key.",
        "description": "By `Id`, not by student number. A capture client holding a card UID wants\r\n`GET /students/by-card/{cardUid}` instead — the UID <em>is</em> the student number, so that\r\nroute is the one that avoids a lookup round trip.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The student's primary key.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The student.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          },
          "404": {
            "description": "No such student.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "summary": "§6.2 `PUT /students/{id}`. A full replacement of the student's own editable fields.\r\n            \r\n\r\n\r\nSending back `course`, `yearLevel` or `section` — which\r\n`GET /students/{id}` returns — is a 400 with `code: \"FieldIsDerived\"`, not a silent\r\ndrop. See EAMS.Application.Abstractions.StudentWriteOutcome.FieldIsDerived.\r\n",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentWriteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentWriteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "summary": "§6.2 `DELETE /students/{id}` — soft (§4.3 `IsDeleted`).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/students/{id}/cards/{cardId}": {
      "delete": {
        "tags": [
          "Students"
        ],
        "summary": "§6.2 `DELETE /students/{id}/cards/{cardId}` — <b>deactivate</b>. The row survives, because\r\nADR-001 D-3's whole point is that a reissued card carries the same REGNO and a past tap must\r\nkeep resolving to the physical card that produced it.\r\n            \r\n\r\n\r\n204 whether or not the card was still active — the postcondition holds either way, so a retry is\r\nsafe. A card that belongs to another student is still a 404: this URL claims it belongs to this\r\none.\r\n",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "DeviceKey": {
        "type": "apiKey",
        "description": "The kiosk / capture credential, scoped to `attendance.capture` and nothing else.\n\nSend the **entire header value including the scheme word**:\n\n```\nAuthorization: DeviceKey eams_dk_<keyId>_<secret>\n```\n\nIssued once by `POST /api/v1/devices` and never retrievable afterwards — the server stores\nonly a hash. Store it in the platform keystore (`expo-secure-store` / Android Keystore),\nnever in `AsyncStorage`. Lose it and rotate with `POST /api/v1/devices/{id}/regenerate-key`.\n\nCase is significant and lower-case is the only accepted form for the token itself.\n\nFailures: `401 DeviceKeyMissing` / `DeviceKeyMalformed` / `DeviceKeyInvalid`,\n`403 DeviceKeyRevoked` / `DeviceInactive`, `429 RateLimited` (honour `Retry-After`).\n`DeviceKeyInvalid` deliberately does not distinguish an unknown key id from a wrong secret.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}